# api/docs/swagger.yaml
definitions:
...
  ek_entity_service.GetProperties:
    properties:
      order:
        type: integer
      property_id:
        type: string
    type: object
  ek_entity_service.GroupPropertySwag:
    properties:
      description:
        example: I have no idea
        type: string
      name:
        type: string
      organization:
        $ref: '#/definitions/ek_entity_service.OrganizationGet'
        type: object
      properties:
        items:
          $ref: '#/definitions/ek_entity_service.GetProperties'
        type: array
      read_statuses:
        example:
        - 60dd9c0a4472a2aaa970304e
        items:
          type: string
        type: array
      status:
        type: boolean
      step:
        type: integer
      type:
        type: integer
      write_statuses:
        example:
        - 60dd9c1a729317449b1ada03
        items:
          type: string
        type: array
    required:
    - description
    - name
    - properties
    - read_statuses
    - status
    - step
    - type
    - write_statuses
    type: object
  ek_entity_service.OrganizationGet:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  ek_variables.CreateResponse:
    properties:
      id:
        type: string
    type: object
...
  ek_variables.FailureResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entity_service.GetAllGroupPropertiesByTypeResponse:
    properties:
      count:
        type: integer
      group_properties:
        items:
          $ref: '#/definitions/entity_service.GroupProperty'
        type: array
    type: object
  entity_service.GetAllGroupPropertiesResponse:
    properties:
      count:
        type: integer
      group_properties:
        items:
          $ref: '#/definitions/entity_service.GetAllGroupProperty'
        type: array
    type: object
  entity_service.GetAllGroupProperty:
    properties:
      category:
        type: integer
      description:
        type: string
      id:
        type: string
      name:
        type: string
      organization:
        $ref: '#/definitions/entity_service.OrganizationModel'
        type: object
      properties:
        items:
          $ref: '#/definitions/entity_service.GetProperties'
        type: array
      read_statuses:
        items:
          type: string
        type: array
      status:
        type: boolean
      step:
        type: integer
      type:
        type: integer
      write_statuses:
        items:
          type: string
        type: array
    type: object
  entity_service.GetAllGroupPropertyByStatusID:
    properties:
      category:
        type: integer
      description:
        type: string
      id:
        type: string
      is_disable:
        type: boolean
      name:
        type: string
      organization:
        $ref: '#/definitions/entity_service.OrganizationModel'
        type: object
      properties:
        items:
          $ref: '#/definitions/entity_service.Property'
        type: array
      status:
        type: boolean
      step:
        type: integer
      type:
        type: integer
    type: object
  entity_service.GetAllGroupPropertyByStatusIdResponse:
    properties:
      group_properties:
        items:
          $ref: '#/definitions/entity_service.GetAllGroupPropertyByStatusID'
        type: array
    type: object
...
  entity_service.GetProperties:
    properties:
      order:
        type: integer
      property_id:
        type: string
    type: object
...
  entity_service.GroupProperty:
    properties:
      category:
        type: integer
      description:
        type: string
      id:
        type: string
      name:
        type: string
      organization:
        $ref: '#/definitions/entity_service.OrganizationModel'
        type: object
      properties:
        items:
          $ref: '#/definitions/entity_service.Property'
        type: array
      read_statuses:
        items:
          $ref: '#/definitions/entity_service.Status'
        type: array
      status:
        type: boolean
      step:
        type: integer
      type:
        type: integer
      write_statuses:
        items:
          $ref: '#/definitions/entity_service.Status'
        type: array
    type: object
  entity_service.OrganizationModel:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  entity_service.Property:
    properties:
      collection_name:
        type: string
      description:
        type: string
      id:
        type: string
      is_required:
        type: boolean
      label:
        type: string
      name:
        type: string
      placeholder:
        type: string
      property_options:
        items:
          $ref: '#/definitions/entity_service.PropertyOption'
        type: array
      status:
        type: boolean
      type:
        type: string
      validation:
        type: string
      with_confirmation:
        type: boolean
    type: object
  entity_service.PropertyOption:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  entity_service.Status:
    properties:
      code:
        type: string
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      parent_status_id:
        type: string
      ru_name:
        type: string
      type_code:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /v1/group-property:
    get:
      consumes:
      - application/json
      description: API for getting all group properties
      parameters:
      - description: search
        in: query
        name: search
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity_service.GetAllGroupPropertiesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Getting All Group Properties
      tags:
      - group_property
    post:
      consumes:
      - application/json
      description: API for creating group_property
      parameters:
      - description: group property
        in: body
        name: group_property
        required: true
        schema:
          $ref: '#/definitions/ek_entity_service.GroupPropertySwag'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ek_variables.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Create group_property
      tags:
      - group_property
  /v1/group-property-status/{status_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: status_id
        in: path
        name: status_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity_service.GetAllGroupPropertyByStatusIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Get group property
      tags:
      - group_property
  /v1/group-property-type:
    get:
      consumes:
      - application/json
      description: API for getting all group properties by type
      parameters:
      - description: step
        in: query
        name: step
        required: true
        type: integer
      - description: type
        in: query
        name: type
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity_service.GetAllGroupPropertiesByTypeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Getting All Group Properties By Type
      tags:
      - group_property
  /v1/group-property/{group_property_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: group_property_id
        in: path
        name: group_property_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity_service.GroupProperty'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Get group property
      tags:
      - group_property
    put:
      consumes:
      - application/json
      description: API for updating group property
      parameters:
      - description: group_property_id
        in: path
        name: group_property_id
        required: true
        type: string
      - description: group property
        in: body
        name: property
        required: true
        schema:
          $ref: '#/definitions/ek_entity_service.GroupPropertySwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ek_variables.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Update Group Property
      tags:
      - group_property
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"